Changed API to output floats, use the helper for signed 16bit stereo

+++ actionamics.h
@@ -7,7 +7,7 @@
 // Public API:
 //   struct actionamics_state *actionamics_init(void *data, uint32_t len, int32_t sample_rate);
 //   void actionamics_free(struct actionamics_state *s);
-//   void actionamics_get_audio(struct actionamics_state *s, int16_t *output, int32_t frames);
+//   void actionamics_get_audio(struct actionamics_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -1524,7 +1524,7 @@ static void actionamics_free(struct actionamics_state *s) {
 }
 
 // [=]===^=[ actionamics_get_audio ]===============================================================[=]
-static void actionamics_get_audio(struct actionamics_state *s, int16_t *output, int32_t frames) {
+static void actionamics_get_audio(struct actionamics_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -1552,7 +1552,7 @@ static void actionamics_api_free(void *state) {
 }
 
 // [=]===^=[ actionamics_api_get_audio ]===========================================================[=]
-static void actionamics_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void actionamics_api_get_audio(void *state, float *output, int32_t frames) {
 	actionamics_get_audio((struct actionamics_state *)state, output, frames);
 }
