+++ amosmusicbank.h
@@ -7,7 +7,7 @@
 // Public API:
 //   struct amosmusicbank_state *amosmusicbank_init(void *data, uint32_t len, int32_t sample_rate);
 //   void amosmusicbank_free(struct amosmusicbank_state *s);
-//   void amosmusicbank_get_audio(struct amosmusicbank_state *s, int16_t *output, int32_t frames);
+//   void amosmusicbank_get_audio(struct amosmusicbank_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -983,7 +983,7 @@ static void amosmusicbank_free(struct amosmusicbank_state *s) {
 }
 
 // [=]===^=[ amosmusicbank_get_audio ]============================================================[=]
-static void amosmusicbank_get_audio(struct amosmusicbank_state *s, int16_t *output, int32_t frames) {
+static void amosmusicbank_get_audio(struct amosmusicbank_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -1011,7 +1011,7 @@ static void amosmusicbank_api_free(void *state) {
 }
 
 // [=]===^=[ amosmusicbank_api_get_audio ]========================================================[=]
-static void amosmusicbank_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void amosmusicbank_api_get_audio(void *state, float *output, int32_t frames) {
 	amosmusicbank_get_audio((struct amosmusicbank_state *)state, output, frames);
 }
