+++ fred.h
@@ -7,7 +7,7 @@
 // Public API:
 //   struct fred_state *fred_init(void *data, uint32_t len, int32_t sample_rate);
 //   void fred_free(struct fred_state *s);
-//   void fred_get_audio(struct fred_state *s, int16_t *output, int32_t frames);
+//   void fred_get_audio(struct fred_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -895,7 +895,7 @@ static void fred_free(struct fred_state *s) {
 }
 
 // [=]===^=[ fred_get_audio ]=====================================================================[=]
-static void fred_get_audio(struct fred_state *s, int16_t *output, int32_t frames) {
+static void fred_get_audio(struct fred_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -923,7 +923,7 @@ static void fred_api_free(void *state) {
 }
 
 // [=]===^=[ fred_api_get_audio ]=================================================================[=]
-static void fred_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void fred_api_get_audio(void *state, float *output, int32_t frames) {
 	fred_get_audio((struct fred_state *)state, output, frames);
 }
