+++ jamcracker.h
@@ -7,7 +7,7 @@
 // Public API:
 //   struct jamcracker_state *jamcracker_init(void *data, uint32_t len, int32_t sample_rate);
 //   void jamcracker_free(struct jamcracker_state *s);
-//   void jamcracker_get_audio(struct jamcracker_state *s, int16_t *output, int32_t frames);
+//   void jamcracker_get_audio(struct jamcracker_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -681,7 +681,7 @@ static void jamcracker_free(struct jamcracker_state *s) {
 }
 
 // [=]===^=[ jamcracker_get_audio ]===============================================================[=]
-static void jamcracker_get_audio(struct jamcracker_state *s, int16_t *output, int32_t frames) {
+static void jamcracker_get_audio(struct jamcracker_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -709,7 +709,7 @@ static void jamcracker_api_free(void *state) {
 }
 
 // [=]===^=[ jamcracker_api_get_audio ]===========================================================[=]
-static void jamcracker_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void jamcracker_api_get_audio(void *state, float *output, int32_t frames) {
 	jamcracker_get_audio((struct jamcracker_state *)state, output, frames);
 }
